+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- encode error as header  [#10](https://github.com/xmidt-org/argus/pull/10)
+- add ability to disable pullInterval for chrysom  [#10](https://github.com/xmidt-org/argus/pull/10)
+- remove authHeader from logging [#10](https://github.com/xmidt-org/argus/pull/10)
+- update response header and status code for invalided requests [#10](https://github.com/xmidt-org/argus/pull/10)
 
 ## [v0.3.2]
 - fix missing conf folder